+++ src/app/registration/registration_form_validates/password_is_not_matches_validate.ts
@@ -1,0 +1,11 @@
+import { FormGroup} from '@angular/forms';
+
+export function ValidatePassword(form: FormGroup) {
+  const password = form.controls.pwd.value;
+  const confirmpwd = form.controls.confirmpwd.value;
+  if (password !== confirmpwd) {
+    form.controls.confirmpwd.setErrors( {MatchPassword: true} )
+  } else {
+    return null
+  }
+}